Create Motion and Contact Trigger
If my Selected (Virtual) Switch is True, then this SmartApp turns selected Dimming Lights (including Hues) ON to past level using either contact or motion sensors as the event trigger.

If no motion for X minutes, then turns selected Lights/Switches OFF -- Unless, the switch was already on prior to motion, in which case it stays on.  If the switch is toggled while the no-motion timer is running, the timer is canceled.